Move observer abstractions to package ...internal.observer
